refactor: update import paths to use '@/'
