--- src/components/controls/bpm/index.ts
@@ -1,1 +1,0 @@
-export * from "./ControlBPM";
