--- src/components/controls/buttons/index.ts
@@ -1,1 +1,0 @@
-export * from "./ControlButtons";
