--- src/components/controls/index.ts
@@ -1,3 +1,0 @@
-export * from "./bpm";
-export * from "./buttons";
-export * from "./ControlPanel";
