+++ src/components/index.ts
@@ -1,4 +1,3 @@
 export * from "./grid";
-export * from "./controls";
 export * from "./legend";
 export * from "./sequencer";
